Move requiredPropCheck util to ext file
